bittrex: convert to TS
