+++ src/ClientOptions.ts
@@ -2,4 +2,5 @@ export type ClientOptions = {
     wssPath?: string;
     watcherMs?: number;
     l2UpdateDepth?: number;
+    throttleL2Snapshot?: number;
 };
